pyproject.toml
[tool.poetry]
name = "minecraft-telegram-server"
version = "0.1.0"
description = ""
authors = ["Ruurd Bijlsma <ruurd@bijlsma.dev>"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.format]
line-ending = "lf"